Core: Added the class name to the info returned by BaseClass::GetInfo().
It uses some compiler trickery because GCC and clang return both a mangled name when calling typeid(Type).name(), while MSVC returns a non-mangled name, but it has the `class` or `struct` keyword in front, so we need to also remove that.

On GCC and clang, a compiler specific function is used to demangle the name, on MSVC, we just remove the first word of the string.